requirements.txt
gql==3.4.0
requests==2.28.1
requests-toolbelt==1.0.0
Jinja2==3.1.2
python-dateutil==2.8.2
python-dotenv==1.0.0 